requirements.txt
absl-py==2.2.2
alembic==1.13.1
amqp==5.3.1
ansicon==1.89.0
anyio==4.9.0
APScheduler==3.11.0
arrow==1.3.0
asgiref==3.8.1
backports.tarfile==1.2.0
bcrypt==4.3.0
billiard==4.2.1
blessed==1.21.0
blinker==1.9.0
build==1.2.2.post1
CacheControl==0.14.2
cachetools==5.5.2
celery==5.5.2
certifi==2025.1.31
cffi==1.17.1
channels==4.2.2
chardet==5.2.0
charset-normalizer==3.4.1
cleo==2.1.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cmdstanpy==1.2.5
colorama==0.4.6
contourpy==1.3.2
crashtest==0.4.1
cron-descriptor==1.4.5
cryptography==44.0.2
cycler==0.12.1
defusedxml==0.7.1
distlib==0.3.9
dj-config-url==0.1.1
dj-database-url==2.1.0
Django==5.2
django-allauth==0.61.1
django-autocomplete-light==3.12.1
django-celery-beat==2.8.0
django-cors-headers==4.7.0
django-crispy-forms==2.4
django-crontab==0.7.1
django-dbbackup==4.2.1
django-defender==0.9.8
django-filter==25.1
django-otp==1.6.0
django-picklefield==3.3
django-q2==1.8.0
django-redis==5.4.0
django-storages==1.14.6
django-timezone-field==7.1
django_celery_results==2.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0
dotenv==0.9.9
dulwich==0.22.8
fastjsonschema==2.21.1
filelock==3.18.0
findpython==0.6.3
Flask==3.0.2
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
fonttools==4.58.0
google-api-core==2.24.2
google-api-python-client==2.169.0
google-auth==2.40.1
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-cloud-core==2.4.3
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
greenlet==3.2.2
gunicorn==21.2.0
h11==0.14.0
holidays==0.72
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
immutabledict==4.2.1
importlib_metadata==8.6.1
importlib_resources==6.5.2
iniconfig==2.1.0
installer==0.7.0
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
Jinja2==3.1.6
jinxed==1.3.0
joblib==1.5.0
keyring==25.6.0
kiwisolver==1.4.8
kombu==5.5.3
Mako==1.3.10
MarkupSafe==3.0.2
matplotlib==3.10.3
more-itertools==10.6.0
msgpack==1.1.0
numpy==2.2.5
oauthlib==3.2.2
ortools==9.12.4544
packaging==24.2
pandas==2.2.3
pbs-installer==2025.3.17
pillow==10.2.0
pkginfo==1.12.1.2
platformdirs==4.3.7
pluggy==1.5.0
poetry==2.1.3
poetry-core==2.1.3
prompt_toolkit==3.0.51
prophet==1.1.6
proto-plus==1.26.1
protobuf==5.29.4
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.9.0
pyotp==2.9.0
pyparsing==3.2.3
pyproject_hooks==1.2.0
pytest==8.0.0
pytest-flask==1.3.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2025.2
pywin32-ctypes==0.2.3
qrcode==8.1
RapidFuzz==3.13.0
redis==3.5.3
reportlab==4.3.1
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rsa==4.9.1
scikit-learn==1.6.1
scipy==1.15.3
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
sqlparse==0.5.3
stanio==0.5.1
threadpoolctl==3.6.0
tomlkit==0.13.2
tqdm==4.67.1
trove-classifiers==2025.3.19.19
types-python-dateutil==2.9.0.20250516
typing_extensions==4.13.1
tzdata==2025.1
tzlocal==5.3.1
uritemplate==4.1.1
urllib3==2.3.0
vine==5.1.0
virtualenv==20.30.0
waitress==2.1.2
wcwidth==0.2.13
Werkzeug==3.0.1
whitenoise==6.6.0
zipp==3.21.0
zstandard==0.23.0
